CCCT-2460 PersonalID Profile Screen Phase Five
[AI] Extracted the text-change listener boilerplate into a reusable EditText.onTextChanged extension.
